+++ src/main/java/insane96mcp/insanesurvivaloverhaul/setup/ModIds.java
@@ -2,4 +2,5 @@ package insane96mcp.insanesurvivaloverhaul.setup;
 
 public class ModIds {
     public static final String RUNE_ENCHANTING = "runeenchanting";
+    public static final String EXPERIENCE_TWEAKS = "experiencetweaks";
 }